requirements.txt
Package                Version
---------------------- --------
click                  8.1.3
et-xmlfile             1.1.0
Flask                  2.2.2
greenlet               1.1.3
importlib-metadata     4.12.0
itsdangerous           2.1.2
Jinja2                 3.1.2
MarkupSafe             2.1.1
mysql-connector-python 8.0.30
mysqlclient            2.1.1
numpy                  1.23.3
openpyxl               3.0.10
pandas                 1.5.0
pip                    22.2.2
protobuf               3.20.1
psycopg2               2.9.3
python-dateutil        2.8.2
python-dotenv          0.21.0
pytz                   2022.2.1
setuptools             41.2.0
six                    1.16.0
SQLAlchemy             1.4.41
Werkzeug               2.2.2
zipp                   3.8.1
